Shifted helper functions `hasErrorText` and `hasNoErrorText` to central utility testing file
