+++ testing/BUILD.bazel
@@ -24,6 +24,7 @@ kt_android_library(
         artifact("org.robolectric:robolectric"),
         artifact("org.jetbrains.kotlin:kotlin-reflect"),
         artifact("androidx.lifecycle:lifecycle-livedata-ktx"),
+        artifact("com.google.android.material:material:1.2.0-alpha02"),
         artifact("com.google.truth:truth"),
         artifact("androidx.test.ext:junit"),
         artifact("org.mockito:mockito-core"),
