+++ testing/build.gradle
@@ -52,6 +52,7 @@ dependencies {
       'androidx.lifecycle:lifecycle-livedata-ktx:2.2.0-alpha03',
       'androidx.test.espresso:espresso-core:3.2.0',
       'androidx.test:runner:1.2.0',
+      'com.google.android.material:material:1.2.0-alpha02',
       'com.google.dagger:dagger:2.24',
       'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.2.2',
       'org.robolectric:robolectric:4.3',
